// tts-learning-be/src/common/types/error-code.ts
export enum ErrorCode {
  // Base
  SUCCESS = 0,
  INTERNAL_SERVER_ERROR = 1000,
  HTTP_NOT_FOUND = 1001,
  HTTP_BAD_REQUEST = 1002,
  HTTP_UNAUTHORIZED = 1003,
  HTTP_FORBIDDEN = 1004,
  HTTP_METHOD_NOT_ALLOWED = 1005,
  HTTP_CONFLICT = 1006,
  HTTP_UNPROCESSABLE_ENTITY = 1007,
  HTTP_TOO_MANY_REQUESTS = 1008,
  HTTP_INTERNAL_SERVER_ERROR = 1009,

  //- Module

  //- Authz
  LOGIN_INVALID = 1100,
  AUTHZ_NOT_FOUND = 1101,
  PASSWORD_INVALID = 1102,

  //- User
  USER_NOT_FOUND = 1200,

  //- Jwt
  INVALID_TOKEN = 1300,
}